Standardize MediFinder preview GIFs width in README
This commit standardizes the display of all MediFinder project preview
GIFs in the README.md file. The width for each GIF (MediFinder1.gif,
MediFinder2.gif, and MediFinder3.gif) has been changed from 33% to a fixed
width of 400 pixels. This adjustment ensures that the preview images are
consistently displayed at the same size, improving the visual layout when
viewing the README on platforms that do not support percentage-based widths
for images.

- Change image width from '33%' to '400px' for uniformity